add ReionRestController, MessageRestController. fix bugs.

+++ src/main/java/by/afanasyeu/avtoxam/dao/mappers/MessageMapper.java
@@ -122,7 +122,7 @@ public interface MessageMapper {
                     "ORDER BY m.id LIMIT #{first}, #{count}"
     )
     @Results()
-    List<MessageDTO> getLikedMessageFromInterval(@Param("first") Long first, @Param("count") Long count,
+    List<MessageDTO> getLikedMessageFromInterval(@Param("first") Long first, @Param("count") Integer count,
                                                  @Param("userId") Long userId);
 
     //обратный порядок
